add check ip into aga-scan-ip.service add update AGA.conf to live-config

+++ etc/webc/aga-scan-ip.sh
@@ -16,22 +16,6 @@ AGA_ID=1
 AGA_mode="kiosk"
 
 
-AGA_update_cfg() {
-  # if writable
-  if touch ${live_image}
-    then
-      # Cache cmdline in case subsequent boots can't reach $config_url
-      cp "$AGA_cfg_file" "$AGA_cfg_file_cached"
-      logs "CONFIG: cached $(md5sum $AGA_cfg_file_cached) $(tr '\n' ' ' < $AGA_cfg_file_cached)"
-      # Kicks off an upgrade
-      mkfifo $upgrade_pipe
-    else
-     # ${live_image} could not be made writable (e.g. live version: booting
-     # from an iso fs), so just use the new config downloaded
-     # and skip all the other stuff below
-     logs "CONFIG: Not a writable boot medium. Could not cache configuration nor upgrade."
-  fi
-}
 
 AGA_has_eth0()
 {
@@ -55,23 +39,8 @@ AGA_has_network()
 # copy that to /etc/webc, so we can compare the new version with
 # it to detect changes and/or use it in case the new download
 # fails.
-if test -s "$AGA_cfg_file_cached"
-  then
-    cp "$AGA_cfg_file_cached" "$AGA_cfg_file"
-    logs "CONFIG: Applied cache $(md5sum $AGA_cfg_file)"
-  else
-    touch "$AGA_cfg_file"
-    logs "CONFIG: No cache"
-fi
-if test -s "$AGA_mod_file_cached"
-  then
-    cp "$AGA_mod_file_cached" "$AGA_mod_file"
-    logs "CONFIG: Applied cache $(md5sum $AGA_mod_file)"
-  else
-    touch "$AGA_mod_file"
-    logs "CONFIG: No cache"
-fi
-
+test -s "$AGA_cfg_file_cached" && cp "$AGA_cfg_file_cached" "$AGA_cfg_file"
+test -s "$AGA_mod_file_cached" && cp "$AGA_mod_file_cached" "$AGA_mod_file"
 . "$AGA_cfg_file"
 . "$AGA_mod_file"
 
@@ -87,7 +56,6 @@ if AGA_has_network; then
   echo AGA_net=\""unknown-net"\" > ${AGA_cfg_file}
   echo AGA_shop_id=\""unknown-shop"\" >> ${AGA_cfg_file}
   chmod 644 ${AGA_cfg_file}
-  AGA_update_cfg
   exit
 fi
 
@@ -139,7 +107,6 @@ iface eth0 inet static
 EOF
     echo nameserver 172.$AGA_net.$AGA_shop_id.1 > ${AGA_resolv_file}
     ifup eth0
-    AGA_update_cfg
     exit
   fi
 fi
@@ -188,7 +155,6 @@ iface eth0 inet static
 EOF
         echo nameserver 172.$AGA_net.$AGA_shop_id.1 > $AGA_resolv_file
         ifup eth0
-        AGA_update_cfg
         exit
       fi
     done
